Let's start by making the stack pointer 64 bit and nobody will get hurt.

+++ lib/libc_r/arch/alpha/uthread_machdep.h
@@ -1,6 +1,6 @@
-/*	$OpenBSD: uthread_machdep.h,v 1.7 2000/10/05 13:27:55 d Exp $	*/
+/*	$OpenBSD: uthread_machdep.h,v 1.8 2002/01/04 14:46:18 art Exp $	*/
 /* David Leonard, <d@csee.uq.edu.au>. Public domain. */
 
 struct _machdep_state {
-	int	sp;
+	u_int64_t	sp;
 };
